gradio/requirements.txt
gradio==3.2
python-dotenv==1.0.0